Mail when curl POST indexing data errors
See https://github.com/hbz/mabxml-elasticsearch/issues/52.

+++ cron.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
-set -euo pipefail # See http://redsymbol.net/articles/unofficial-bash-strict-mode/
+# explicitly without "-e" for it should not exit immediately when failed but write a mail
+set -uo pipefail # See http://redsymbol.net/articles/unofficial-bash-strict-mode/
 IFS=$'\n\t'
 
 # Execute via crontab by hduser@weywot1:
@@ -9,6 +10,8 @@ IFS=$'\n\t'
 TARGET_PATH=/data/DE-605/mabxml
 # Determine the latest update file and store it locally:
 DATE=$(date "+%Y%m%d")
+
+RECIPIENT=lobid-admin
 
 # Use "brace extension" as we don't know the appendix of the basedump
 if [ $1 == "basedump" ]; then
@@ -20,5 +23,12 @@ fi
 # Run the transformation with the latest file (and possibly unprocessed previous files):
 curl --fail -XPOST "http://localhost:7300/hbz01/transform?dir=$TARGET_PATH&suffix=gz&cluster=weywot&hostname=10.9.0.12&index=hbz01" >> logs/processMabxml.sh.$DATE-weywot.log 2>&1
 
+# if check ok, index to productive instance:
+if [  $? -ne 0 ]; then
+    mail -s "Alert MAB-XML Clobs indexing to hbz01" "$RECIPIENT@hbz-nrw.de" << EOF
+"error when indexing MAB-XML Clobs to index hbz01 :("
+EOF
+else
 # Clean up
-rm $TARGET_PATH/*.gz
+  rm $TARGET_PATH/*.gz
+fi
